Refactor the services
