+++ src/services/ProviderService.ts
@@ -1,19 +1,16 @@
 import { HTTP_STATUS_CODE } from "@/constants";
 import { hashPassword } from "@/lib/passwordHashing";
 import UserParser from "@/parsers/UserParser";
-import { AuthRepository } from "@/repository/AuthRepository";
-import UserRepository from "@/repository/UserRepository";
+import { ProviderRepository } from "@/repository/ProviderRepository";
 import HTTPError from "@/utils/error/HTTPError";
 import { FastifyReply, FastifyRequest } from "fastify";
 import { handleServiceError } from ".";
 
 export class ProviderService {
-  private providerRepository: UserRepository | undefined;
-  private authRepository: AuthRepository | undefined;
+  private providerRepository = new ProviderRepository();
   private readonly parser = new UserParser();
 
   async createProvider(request: FastifyRequest, reply: FastifyReply) {
-    this.providerRepository = new UserRepository();
     try {
       const { password, email, phoneNumber, ...userData } =
         this.parser.parseBodyForUserCreation(request);
@@ -31,7 +28,7 @@ export class ProviderService {
           "An user with this phone number already exists"
         );
       const hashedPassword = await hashPassword(password);
-      const createdClient = await this.providerRepository.createProvier({
+      const createdClient = await this.providerRepository.createProvider({
         auth: {
           email,
           password: hashedPassword,
@@ -50,17 +47,14 @@ export class ProviderService {
     request: FastifyRequest,
     reply: FastifyReply
   ) {
-    this.providerRepository = new UserRepository();
-    this.authRepository = new AuthRepository();
     try {
       const { url } = this.parser.parseBodyForImageUpdate(request);
       const { email } = request.user;
-      const user = await this.authRepository.getByEmail(email);
+      const user = await this.providerRepository.getByEmail(email);
       if (!user)
         throw new HTTPError(HTTP_STATUS_CODE.NOT_FOUND, "User not found");
-      const { id: userId } = user.User!;
       const updatedProvider =
-        await this.providerRepository.updateProviderAvatarImageURL(url, userId);
+        await this.providerRepository.updateUserAvatarImageURL(url, user.id);
       const { avatarImageURL } = updatedProvider;
       reply.code(HTTP_STATUS_CODE.OK).send({
         avatarImageURL,
